moved stuff to cpp files where possible, fixed some memory issues
